Address secruity concerns with hackernews example
This change prevents a carefully crafted hackernews story from executing arbitary JavaScript in Craft when the hackernews example extension is active.

It:

- Introduces proper encodeURIComponent encoding on untrusted strings
- Introduces proper entity escaping on untrusted strings
- Stops embedded untrusted strings directly into JavaScript source

While, in this particular case, the attack vector is hard to execute (requiring placing an malicious story in the top ten of hackernews) as a demo of "best practices" this example needs improvement.